+++ packages/builder/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @d-dash/builder
 
+## 0.0.2
+
+### Patch Changes
+
+- replace stat with gauge
+- Updated dependencies
+  - @d-dash/core@0.3.1
+
 ## 0.0.1
 
 - First release